oth: update structure for tests
